% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_word_info.R
\name{assign_word_info}
\alias{assign_word_info}
\title{Assign Word Information to Fixations}
\usage{
assign_word_info(fixations, word_boundaries, sentence_left_x_boundary = 125)
}
\arguments{
\item{fixations}{A data frame containing fixation data, usually from the output of \code{read_asc_file}}

\item{word_boundaries}{A data frame containing word boundary information, usually from the output of \code{read_asc_file}.}

\item{sentence_left_x_boundary}{X-coordinate boundary indicating the left edge of the sentence on the screen.}
}
\value{
A modified data frame of fixations with added columns for word number and word text for each fixation.
}
\description{
This function adds information about the fixated word to each fixation in a data frame of fixations based on word boundaries. It determines the word number and word text that was fixated during each fixation.
}
