% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink_overlap_or_near.R
\name{blink_overlap_or_near}
\alias{blink_overlap_or_near}
\title{Evaluate Blink Overlap or Proximity to Fixation}
\usage{
blink_overlap_or_near(
  fixation_start,
  fixation_end,
  blink_start,
  blink_end,
  t_diff_threshold = 100
)
}
\arguments{
\item{fixation_start}{Start timestamp of the fixation in milliseconds.}

\item{fixation_end}{End timestamp of the fixation in milliseconds.}

\item{blink_start}{Start timestamp of the blink in milliseconds.}

\item{blink_end}{End timestamp of the blink in milliseconds.}

\item{t_diff_threshold}{Time difference threshold (in milliseconds) for considering overlap or proximity.}
}
\value{
Logical value indicating whether the fixation and blink overlap or are near each other in time.
}
\description{
This function evaluates whether a single fixation and a single blink are close in time, indicating an overlap or proximity between them based on a specified time difference threshold.
}
\examples{
# Check if a fixation and a blink are close
fixation_near_blink <- blink_overlap_or_near(100, 180, 190, 220, t_diff_threshold = 100)

}
