% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fixation_time_measures.R
\name{calculate_fixation_time_measures}
\alias{calculate_fixation_time_measures}
\title{Calculate Fixation Time Measures}
\usage{
calculate_fixation_time_measures(fixations, words)
}
\arguments{
\item{fixations}{A data frame containing fixation data.}

\item{words}{A data frame containing word information.}
}
\value{
A data frame containing fixation time measures for each word. The resulting data frame includes the following columns:

\describe{
\item{word_nr}{Word number associated with the fixation.}
\item{ffd}{First Fixation Duration (in milliseconds) - Duration of the first fixation on the word.}
\item{gd}{Gaze Duration (in milliseconds) - Total duration of fixations (first fixation and refixations) on the word during first-pass.}
\item{sfd}{Single Fixation Duration (in milliseconds) - If the first fixation duration is equal to gaze duration, then SFD is the same as FFD, otherwise NA.}
\item{gopast}{Go-past time (in milliseconds) - Total time spent from first entering word from the left to leaving word to the right, including any regressions. Also known as regression path duration.}
\item{tvt}{Total Viewing Time (in milliseconds) - Total time spent on the word during the entire trial.}
\item{word}{Word text associated with the word number.}
}
}
\description{
This function calculates fixation time measures based on a data frame of fixations and associated word information. The calculated measures include measures related to first pass reading, gaze duration, and total viewing time for each word.
}
