% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_subject_EK_measures.R
\name{calculate_subject_EK_measures}
\alias{calculate_subject_EK_measures}
\title{Calculate Eye-Tracking Measures for a Subject}
\usage{
calculate_subject_EK_measures(
  subject,
  reduce_sampling_rate_to = NULL,
  downsampling_method = "drop",
  lambda = 6,
  return_trial_fixations = FALSE,
  use_eye = "R"
)
}
\arguments{
\item{subject}{A list containing subject-level data elements, usually created by the \code{read_asc_file} function. The list should include the following elements:
\itemize{
\item \code{filename}: Character string representing the subject's filename or identifier.
\item \code{rate}: Character string specifying the recording sampling rate used in the experiment (e.g., "1000" for 1000 Hz).
\item \code{trial_db}: A data frame containing trial-level information for each trial in the experiment.
\item \code{trial_samples}: A list containing data frames, where each data frame represents eye-tracking samples for a trial.
\item \code{trial_words}: A list containing data frames, where each data frame holds word information for a trial.
\item \code{trial_fixations}: A list containing data frames, where each data frame contains fixation data for a trial (optional, used if \code{return_trial_fixations} is \code{TRUE}).
\item \code{trial_saccades}: A list containing data frames, where each data frame holds saccade information for a trial.
\item \code{trial_blinks}: A list containing data frames, where each data frame represents blink information for a trial.
}

These elements are used to process each trial's eye-tracking data and calculate fixation-related measures for analysis.}

\item{reduce_sampling_rate_to}{Numeric value specifying the target sampling rate after downsampling (Hz). \code{NULL} if no downsampling is desired.}

\item{downsampling_method}{Method used for downsampling. \code{drop} will keep one of every X samples (where X, the downsampling factor, is sampling rate/target sampling rate). \code{average} will average over every X samples.}

\item{lambda}{Lambda parameter for the Engbert & Kliegl (2004) fixation detection algorithm as implemented in the saccades package.}

\item{return_trial_fixations}{Logical indicating whether to return trial-specific fixation data.}

\item{use_eye}{Character indicating which eye to use for analysis ("L" for left, "R" for right).}
}
\value{
A data frame containing fixation time measures for each trial.

A list with fixation data frames for each trial (if \code{return_trial_fixations} is \code{TRUE}).
}
\description{
This function calculates various eye-tracking measures for a subject's experimental trials. It processes each trial's eye-tracking data, performs downsampling if required, detects fixations using the Engbert & Kliegl (2004) algorithm as implemented in the \code{saccades} package, and calculates fixation-related measures such as fixation time measures and other trial-specific information.
}
