% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_downsampling_factor.R
\name{determine_downsampling_factor}
\alias{determine_downsampling_factor}
\title{Determine Downsampling Factor}
\usage{
determine_downsampling_factor(
  original_sampling_rate,
  reduce_sampling_rate_to,
  downsampling_method = "drop"
)
}
\arguments{
\item{original_sampling_rate}{The original sampling rate of the data in Hz.}

\item{reduce_sampling_rate_to}{The downsampling method used, \code{drop} or \code{average}. This only affects the message displayed}
}
\value{
The calculated downsampling factor.
}
\description{
This function calculates the downsampling factor based on the original sampling rate and the target reduced sampling rate.
}
\details{
The downsampling factor is calculated as the ratio of the original sampling rate to the target reduced sampling rate. The function also checks if the target sampling rate is higher than the original sampling rate and provides warnings if necessary. If the original sampling rate is not a multiple of the target sampling rate, the function adjusts the downsampling factor to ensure that the new sampling rate is a multiple of the target rate.
}
\examples{
# Original sampling rate and target reduced sampling rate
original_sampling_rate <- 2000
reduce_sampling_rate_to <- 1000

# Calculate the downsampling factor
downsampling_factor <- determine_downsampling_factor(original_sampling_rate, reduce_sampling_rate_to)

}
