% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_in_dc.R
\name{fix_in_dc}
\alias{fix_in_dc}
\title{Function to extract fixations from a trial}
\usage{
fix_in_dc(
  eyelink_fixations,
  eyelink_blinks,
  trial_words,
  trial_db_for_trial,
  use_eye = "R"
)
}
\arguments{
\item{eyelink_fixations}{Fixations data from the Eyelink}

\item{eyelink_blinks}{Blinks data from the Eyelink}

\item{trial_words}{Word boundaries for the trial}

\item{trial_db_for_trial}{Data for the trial}

\item{use_eye}{Eye to use for the fixations}
}
\value{
fixations_with_word_info Fixations data with word information
}
\description{
This function extracts fixations from a trial, given the trial's word boundaries and the fixations data. It also assigns the word information to each fixation.
}
\examples{
# Example data
eyelink_fixations <- data.frame(eye = c("R", "R", "L"),
                                t_start = c(100, 200, 300),
                                t_end = c(150, 250, 350))
eyelink_blinks <- data.frame(t_start = c(120, 230),
                             t_end = c(130, 240))
trial_words <- data.frame(word_start = c(50, 160, 270),
                          word_end = c(150, 250, 350))
trial_db_for_trial <- data.frame(timestamp_display_on = 100,
                                 timestamp_display_off = 350,
                                 language = "EN",
                                 subject_nr = 1,
                                 sentence_number = 42)

# Run function with example data
fix_in_dc(eyelink_fixations, eyelink_blinks, trial_words, trial_db_for_trial, use_eye = "R")

}
