% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfViolinPlots.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Flat Violin Plot Geom for ggplot2}
\source{
https://github.com/gabrifc/raincloud-shiny
}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified, it
overrides the default mapping at the plot level.}

\item{data}{A data frame. If specified, it overrides the default data at the plot level.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{trim}{Logical, whether to trim the tails of the violins to the range of the data.}

\item{scale}{If "area" (default), all violins have the same area; if "count," areas are scaled proportionally
to the number of observations; if "width," all violins have the same maximum width.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics rather than combining with them.}

\item{...}{Additional arguments passed to the layer.}
}
\value{
A ggplot2 layer for flat violin plots, useful in creating raincloud plots.
}
\description{
This function adds a custom flat violin plot geom to ggplot2. Unlike the traditional
violin plot, this version displays only one side of the density plot, making it
ideal for raincloud plots and other compact visualizations. This geom was inspired by a solution
to issues raised on social media and utilizes modified code from ggplot2's \code{geom_violin}.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, hwy)) +
  geom_flat_violin(trim = FALSE)

}
\references{
Original issue raised on Twitter: \url{https://twitter.com/EamonCaddigan/status/646759751242620928}
ggplot2's \code{geom_violin} source: \url{https://github.com/hadley/ggplot2/blob/master/R/geom-violin.r}
}
