% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomFlatViolin.R
\docType{data}
\name{GeomFlatViolin}
\alias{GeomFlatViolin}
\title{GeomFlatViolin}
\format{
An object of class \code{GeomFlatViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomFlatViolin
}
\value{
A ggproto object that defines the flat violin geometry.
}
\description{
A custom ggplot2 geom that creates flat violins for visualizing the distribution of a variable.
}
\details{
This geometry allows for a flat violin representation, which is particularly useful for
visualizing the density of the data while maintaining a clear view of the underlying data points.
}
\keyword{datasets}
