% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Get a List of Files with a Specified Extension from a Directory}
\usage{
get_files(dir = "", file_ext = ".asc$")
}
\arguments{
\item{dir}{Character string indicating the directory path where the files are located. If not provided, the current working directory is used.}

\item{file_ext}{Character string representing the desired file extension (e.g., ".asc").}
}
\value{
A character vector containing the full paths to the selected files with the specified extension in the directory.
}
\description{
This function retrieves a list of files with a specified file extension from a user-provided directory. It can filter out non-matching files and sort the files based on numeric order within their names if applicable.
}
\examples{
# Retrieve a list of .asc files from the current directory
asc_files <- get_files(file_ext = ".asc$")

# Retrieve a list of .txt files from a specified directory
txt_files <- get_files(dir = "/path/to/directory", file_ext = ".txt")

}
