% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_fixations_close_to_blinks.R
\name{mark_fixations_close_to_blinks}
\alias{mark_fixations_close_to_blinks}
\title{Mark Fixations Close to Blinks}
\usage{
mark_fixations_close_to_blinks(fixations, blinks, t_diff_threshold = 100)
}
\arguments{
\item{fixations}{A data frame containing fixation data.}

\item{blinks}{A data frame containing detected blink data.}

\item{t_diff_threshold}{Time difference threshold (in milliseconds) for identifying fixations close to blinks. How far away does the start or end of a fixation have to be from the start/end of a blink to be marked as close to a blink?}
}
\value{
A modified data frame of fixations with an additional column indicating whether each fixation is close to a blink.
}
\description{
This function takes a data frame of fixations and a data frame of detected blinks and identifies which fixations are close in time to any of the detected blinks based on a specified time difference threshold.
}
