% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fixations.R
\name{merge_fixations}
\alias{merge_fixations}
\title{Merge Fixations in a Data Frame}
\usage{
merge_fixations(
  df,
  max_duration = 80,
  max_distance = 12,
  rate = 1000,
  add_eyelink_correction = TRUE
)
}
\arguments{
\item{df}{A data frame containing fixation information with columns 'eye', 't_start', t_end', 'duration', 'x', 'y', and 'pupil'.}

\item{max_duration}{The maximum duration in milliseconds for a fixation to be considered for merging (default is 200 ms).}

\item{max_distance}{The maximum distance in pixels between fixations for them to be considered for merging (default is 50 pixels).}

\item{rate}{The sampling rate of the eye tracker in Hz (default is 1000 Hz).}

\item{add_eyelink_correction}{A logical indicating whether to add a correction for the}
}
\value{
A modified data frame with merged fixations and a new column 'merged' indicating
whether a fixation was merged.
}
\description{
This function merges fixations with a duration of less than \code{max_duration}
milliseconds with either the previous or the subsequent fixation, as long as that
fixation is less than \code{max_distance} pixels away. If there are two possibilities for merging,
the previous fixation is chosen.
}
\examples{
fixations <- data.frame(
  eye = c("R", "R", "R", "R", "R"),
  t_start = c(5716010, 5716442, 5716566, 5717018, 5717202),
  t_end = c(5716378, 5716550, 5716986, 5717174, 5717398),
  duration = c(372, 112, 424, 160, 200),
  x = c(962, 202, 160, 294, 231),
  y = c(539, 554, 538, 555, 553),
  pupil = c(649, 651, 630, 622, 612)
)
merged_fixations <- merge_fixations(fixations)
print(merged_fixations)

}
