% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trial.R
\name{plot_trial}
\alias{plot_trial}
\title{Visualize Eye Tracking Experiment Trials}
\usage{
plot_trial(
  subject,
  trial_nr,
  use_eye = "R",
  show_original_samples = TRUE,
  show_downsampled_samples = TRUE,
  downsampling_method = "drop",
  downsampling_rate = 125,
  show_eyelink_fixations = TRUE,
  show_EK_fixations = TRUE,
  show_EK_fixations_downsampled = TRUE,
  show_words = TRUE,
  show_word_boundaries = TRUE,
  x_limits = c(0, 1920),
  y_limits = c(1080, 0),
  lambda = 6,
  lambda_downsampled = lambda,
  smooth.coords = FALSE,
  smooth.saccades = TRUE,
  min_duration_for_elimination = 800,
  max_duration_for_elimination_or_merge = 80,
  max_distance_for_merge = 12,
  letter_width = 12,
  words_font_size = 12,
  df_with_fixed_words = sentences_words_espal
)
}
\arguments{
\item{subject}{The subject data object containing eye tracking and trial information.}

\item{trial_nr}{The trial number to visualize.}

\item{show_original_samples}{Display the original eye tracking samples.}

\item{show_downsampled_samples}{Display the downsampled eye tracking samples.}

\item{downsampling_rate}{The rate at which to downsample the samples (in Hz).}

\item{show_eyelink_fixations}{Display Eyelink fixations.}

\item{show_EK_fixations}{Display fixations determined by the Engbert and Kliegl algorithm from the original samples.}

\item{show_EK_fixations_downsampled}{Display fixations determined by the Engbert and Kliegl algorithm from the downsampled samples.}

\item{show_words}{Display word boundaries and labels.}

\item{show_word_boundaries}{Display word boundaries.}

\item{x_limits}{The x-axis limits for the plot.}

\item{y_limits}{The y-axis limits for the plot.}

\item{lambda}{Smoothing parameter for fixation identification.}

\item{lambda_downsampled}{Smoothing parameter for fixation identification on downsampled samples.}

\item{smooth.coords}{Whether to perform coordinate smoothing on fixation data.}

\item{smooth.saccades}{Whether to perform saccade smoothing on fixation data.}

\item{min_duration_for_elimination}{Minimum duration for eliminating fixations.}

\item{max_duration_for_elimination_or_merge}{Maximum duration for elimination or merge criteria.}

\item{max_distance_for_merge}{Maximum distance for merge criteria.}

\item{letter_width}{The width of a single letter.}
}
\value{
A ggplot object representing the eye tracking data visualization.
}
\description{
This function generates a visualization of eye tracking data for a specific trial of the sampling rate eye tracking experiment.
}
\details{
This function visualizes eye tracking data for a specific trial, including original and downsampled samples, Eyelink fixations, EyeKnows fixations, word boundaries, and labels. It provides options to control which elements to display and allows for customization of plot aesthetics. It may be possible to use this function to plot other eye tracking trials as well.
}
