% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_EK_trial.R
\name{process_EK_trial}
\alias{process_EK_trial}
\title{Process Eye-Tracking Trial Data}
\usage{
process_EK_trial(
  trial_samples,
  trial_words,
  trial_db_for_trial,
  use_eye = "R",
  min_duration_for_elimination = 800,
  max_duration_for_elimination_or_merge = 80,
  max_distance_for_merge = 12,
  sentence_start_x = 125,
  lambda = 6,
  smooth.coords = FALSE,
  smooth.saccades = TRUE,
  downsampling_factor = 1,
  downsampling_method = "drop"
)
}
\arguments{
\item{trial_samples}{A data frame containing eye-tracking samples for the trial.}

\item{trial_words}{A data frame containing word information for the trial.}

\item{trial_db_for_trial}{A data frame containing trial-level information for the trial.}

\item{use_eye}{Character indicating which eye to use for analysis ("L" for left, "R" for right).}

\item{min_duration_for_elimination}{Minimum duration (in milliseconds) for long fixations to be considered for elimination.}

\item{max_duration_for_elimination_or_merge}{Maximum duration (in milliseconds) for short fixations to be considered for elimination or merging.}

\item{max_distance_for_merge}{Maximum distance (in pixels) between fixations for merging.}

\item{sentence_start_x}{X-coordinate (in pixels) indicating the starting position of the sentence on the screen.}

\item{lambda}{Lambda parameter for fixation detection algorithm (see \code{\link[saccades]{detect.fixations}}.}

\item{smooth.coords}{Logical indicating whether to apply coordinate smoothing during fixation detection (see \code{\link[saccades]{detect.fixations}}.}

\item{smooth.saccades}{Logical indicating whether to apply saccade smoothing during fixation detection (see \code{\link[saccades]{detect.fixations}}.}

\item{downsampling_factor}{Downsampling factor for reducing the data. This is for evaluating the performance of the saccade detection algorithm with lower sampling rates.}

\item{downsampling_method}{What method to use to remove samples? Default is "drop": Keep every n-th sample. Alternative: "average": average over n samples (where n is the downsampling factor)}
}
\value{
A data frame containing fixations with added word information.

The returned data frame contains the following columns:

\describe{
\item{trial}{Trial number of the eye-tracking data.}
\item{start}{Timestamp of the fixation start in milliseconds.}
\item{end}{Timestamp of the fixation end in milliseconds.}
\item{x}{X-coordinate of the fixation point on the screen in pixels.}
\item{y}{Y-coordinate of the fixation point on the screen in pixels.}
\item{mad.x}{Median absolute deviation of x-coordinates during the fixation.}
\item{mad.y}{Median absolute deviation of y-coordinates during the fixation.}
\item{peak.vx}{Peak velocity in the x-direction during the fixation.}
\item{peak.vy}{Peak velocity in the y-direction during the fixation.}
\item{dur}{Duration of the fixation in milliseconds.}
\item{event}{Type of eye movement event, e.g., "fixation".}
\item{duration}{Duration of the fixation in milliseconds (duplicate of 'dur' column).}
\item{t_start}{Timestamp of the fixation start in milliseconds (duplicate of 'start' column).}
\item{t_end}{Timestamp of the fixation end in milliseconds (duplicate of 'end' column).}
\item{too_short}{Logical indicating if the fixation duration is too short.}
\item{too_long}{Logical indicating if the fixation duration is too long.}
\item{near_blink}{Logical indicating if the fixation is near a blink region.}
\item{merged}{Logical indicating if the fixation was merged with an adjacent fixation.}
\item{word_nr}{Word number associated with the fixation (-1 for no word).}
\item{word}{Word text fixated (NA if no word fixated).}
\item{sentence_nr}{Sentence number associated with the fixation.}
\item{trial_nr}{Trial number associated with the fixation.}
\item{subject}{Subject identifier for the trial.}
}
}
\description{
This function processes eye-tracking trial data from a reading experiment. It performs various steps including sample filtering, downsampling, fixation detection (using the algorithm by Engbert & Kliegl, 2004, as implemented in the saccades package), fixation merging, and adding word information to the fixations.
}
\examples{
# Load necessary packages
library(dplyr)
library(saccades)

# Process trial data
processed_data <- process_EK_trial(trial_samples, trial_words, trial_db_for_trial)

}
