% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_eyelink_trial.R
\name{process_eyelink_trial}
\alias{process_eyelink_trial}
\title{Processing EyeLink data for a single trial}
\usage{
process_eyelink_trial(
  eyelink_fixations,
  eyelink_blinks,
  trial_words,
  trial_db_for_trial,
  use_eye = "R",
  min_duration_for_elimination = 800,
  max_duration_for_elimination_or_merge = 80,
  max_distance_for_merge = 12,
  sentence_start_x = 51
)
}
\arguments{
\item{eyelink_fixations}{A data frame containing fixation data from the ASC file.}

\item{eyelink_blinks}{A data frame containing blink data from the ASC file.}

\item{trial_words}{A data frame containing word information for the current trial.}

\item{trial_db_for_trial}{A data frame containing trial information for the current trial.}

\item{use_eye}{A character string specifying which eye to use for the analysis. Must be either "L" or "R". Default is "R".}

\item{min_duration_for_elimination}{An integer specifying the minimum duration (in milliseconds) for a fixation to be considered valid. Fixations shorter than this duration will be eliminated. Default is 800 ms.}

\item{max_duration_for_elimination_or_merge}{An integer specifying the maximum duration (in milliseconds) for a fixation to be considered valid. Fixations longer than this duration will be eliminated, and fixations shorter than this duration may be merged with neighboring fixations. Default is 80 ms.}

\item{max_distance_for_merge}{An integer specifying the maximum distance (in pixels) between fixations to be considered for merging. Fixations that are closer than this distance may be merged. Default is 12 px.}

\item{sentence_start_x}{An integer specifying the x-coordinate of the start of the sentence. Default is 51 px.}
}
\value{
A list containing various data frames and lists including trial information, messages, samples, words, fixations, saccades, and blinks.
}
\description{
This function processes EyeLink data for a single trial, including fixations, saccades, and blinks.
}
\examples{
eyelink_fixations <- read_csv("eyelink_fixations.csv")
eyelink_blinks <- read_csv("eyelink_blinks.csv")
trial_words <- read_csv("trial_words.csv")
trial_db_for_trial <- read_csv("trial_db_for_trial.csv")
processed_data <- process_eyelink_trial(eyelink_fixations, eyelink_blinks, trial_words, trial_db_for_trial)

}
\seealso{
merge_fixations

assign_word_info

assign_word_info_eng

mark_fixations_close_to_blinks
}
