% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asc_file.R
\name{read_asc_file}
\alias{read_asc_file}
\title{Read Eyelink ASC File and Extract Data}
\usage{
read_asc_file(data_file, rate = 1000, sentence_start_x = 125, language)
}
\arguments{
\item{data_file}{The path to the ASC data file to be read.}

\item{rate}{The sampling rate of the data.}

\item{sentence_start_x}{The x-coordinate of the sentence start. Default is 125.}
}
\value{
A list containing various data frames and lists including trial information, messages, samples, words, fixations, saccades, and blinks.
}
\description{
This function reads an Eyelink ASC file from an experiment conducted with Opensesame, and extracts various data including raw samples, fixations, blinks, and word information that was written into the ASC file by Opensesame.
}
\examples{
# Provide the path to the ASC file
data_file <- "path/to/your/data.asc"

# Read the ASC file and extract data
subject_data <- read_asc_file(data_file)

}
